Modified beersheba to use the high energy threshold cdst for energy
Due to the masks in DEMO the lowTh of esmeralda needs to be set at
quite low Q. This, combined with the small size of DEMO, produces lots
a non-negligible amount of hits outside the correction map.

To accommodate for this, beersheba has been modified to use the high
threshold cdst to assign the energy and NaN correction factors are
subbed by the closest correction factor in the dst.